src/Impact_AirPollution_Cough_Audio/requirements.txt
streamlit
streamlit-audiorecorder
streamlit_extras
librosa
tensorflow
